refactor: keep ensure and slack plugin entries erasure-identical
The Fargate runner (run-cluster.js) requires both plugins
synchronously; directly-require()d .ts files bypass native type
stripping when require-extension hooks are installed (pirates via
tap's processinfo, inherited by spawned CLI processes through
NODE_OPTIONS). Plugin implementation moves to plugin.ts; index.ts is
a plain re-export with no TypeScript-only syntax.

Also adds a root typecheck script.
